refactor(qwen35): unwrap LayerKind::load and re-inline the weights test fixture
Review pass over the loader split, trimming wrapping that paid no rent:

- LayerKind::load was a pure forwarding hop; the full/linear match now lives
  in TransformerBlock35::load, so a block load is one call, not two.
- weights/fixture.rs was a module serving a single test after the mlp
  arithmetic test retired; the two helpers re-join layers.rs's test module
  and the file is gone.

Evidence (A100-40GB, sm_80, Qwen3.5-4B): cargo check/clippy --all-targets
-D warnings clean; lib tests 99 passed; hf_golden_gate TP1 2/2 and TP2
(--ignored) 2/2.
